Fix CSS Loader with unknown option 'localIdentName'
The modules is valid option property of Loader. Remove 'localIdentName' property from option and assign it to modules property.

The CSS Loader 3.2.0 has been initialised using an options object that does not match the API schema.It specification does not accept 'localIdentName' as property to options.

The modules is valid option property of Loader. The 'localIdentName' is assigned to modules as a property.

More reference related to error:
ValidationError: Invalid options object. CSS Loader has been initialised using an options object that does not match the API schema.
 - options has an unknown property 'localIdentName'. These properties are valid:
   object { url?, import?, modules?, sourceMap?, importLoaders?, localsConvention?, onlyLocals? }